Move betty.plugins.license to betty.licenses
