+++ betty/licenses/all_rights_reserved.py
