+++ betty/licenses/spdx.py
