+++ betty/tests/licenses/test_all_rights_reserved.py
@@ -3,7 +3,7 @@ from typing import override
 import pytest
 
 from betty.license import License
-from betty.plugins.license.all_rights_reserved import AllRightsReserved
+from betty.licenses.all_rights_reserved import AllRightsReserved
 from betty.test_utils.license import LicenseTestBase
