+++ pyproject.toml
@@ -141,10 +141,10 @@ betty = 'betty.console:main_standalone'
 'webpack' = 'betty.plugins.document_provider.webpack:Webpack'
 
 [project.entry-points.'betty.license']
-'all-rights-reserved' = 'betty.plugins.license.all_rights_reserved:AllRightsReserved'
+'all-rights-reserved' = 'betty.licenses.all_rights_reserved:AllRightsReserved'
 'project-configuration' = 'betty.project._plugin_discovery:_discover_licenses'
-'public-domain' = 'betty.plugins.license.public_domain:PublicDomain'
-'spdx' = 'betty.plugins.license.spdx:SpdxLicenseDiscoverer.discover_for'
+'public-domain' = 'betty.licenses.public_domain:PublicDomain'
+'spdx' = 'betty.licenses.spdx:SpdxLicenseDiscoverer.discover_for'
 
 [project.entry-points.'betty.link']
 'betty-documentation' = 'betty.links.betty_documentation:BETTY_DOCUMENTATION'
